refactor(validation): split instruction budget modules
Keep the instruction budget CLI entry point intact while moving constants, glob parsing, and value objects into focused sibling modules.
