+++ scripts/validation/instruction_budget_constants.py
@@ -1,0 +1,15 @@
+"""Constants for instruction budget validation."""
+
+from __future__ import annotations
+
+INSTRUCTIONS_SUBDIR = ".github/instructions"
+INSTRUCTION_GLOB = "*.instructions.md"
+
+# Non-regression ratchet ceilings in bytes, seeded just above current measured
+# values (see module docstring). Lower these as the corpus shrinks.
+DEFAULT_CEILINGS_BYTES: dict[str, int] = {
+    ".py": 220_000,
+    ".cs": 220_000,
+    ".ps1": 220_000,
+    ".md": 83_000,
+}
